backend/requirements.txt
flask
flask_sqlalchemy
flask_login
requests
python-dotenv
